MAJ fonction deleteSauce pour suppression image

+++ controllers/sauce.js
@@ -1,4 +1,5 @@
 const Sauce = require("../models/Sauce");
+const fs = require("fs");
 
 /**Renvoie un tableau de toutes les sauces */
 exports.getAllSauces = (req, res, next) => {
@@ -79,7 +80,23 @@ exports.updateSauce = (req, res, next) => {
 
 /**Supprime la sauce */
 exports.deleteSauce = (req, res, next) => {
-    Sauce.deleteOne({ _id: req.params.id })
-        .then(() => { res.status(200).json({ message: "Votre sauce a bien été supprimée !" }) })
-        .catch((error) => res.status(400).json({ error }));
+    Sauce.findOne({ _id: req.params.id })
+        .then(sauce => {
+            if (sauce.userId != req.auth.userId) {
+                res.status(401).json({ message: "Pas autorisé" });
+            } else {
+                const filename = sauce.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Sauce.deleteOne({ _id: req.params.id })
+                        .then(() => { res.status(200).json({ message: "Votre sauce a bien été supprimée !" }) })
+                        .catch((error) => res.status(400).json({ error }));
+                });
+            }
+        })
+        .catch(error => {
+            res.status(500).json({ error });
+        });
+};
+    
+
 };
